src/requirements.txt
Django==1.11.9
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
gunicorn==19.7.1
